daywatch/requirements.txt
Django==1.6.7
South==0.8.1
psycopg2==2.5.1
PyYAML==3.10
billiard==2.7.3.32
lxml==3.4.4
service-identity==14.0.0

# Remote task execution
amqp==1.0.13
redis==2.8.0
celery==3.0.21

# Django extensions
django-rosetta==0.7.2
django-extensions==1.2.4
django-widget-tweaks==1.3
django-recaptcha==0.0.6
djangorestframework==2.4.3
django-filter==0.7
django-celery==3.0.23
django-libsass
django_compressor
django-appconf
django-bootstrap-pagination==1.5.0

# Deployment
gunicorn==18.0

# For pretty printing
colorama==0.2.7

# For scrapers
scrapy==0.24.4

# Ansible
ansible==1.5
selenium==2.46.1
